+++ packages/core/cli/src/utils/ensureDirectoryExists.ts
